Use rmdir on windows for deletion
shutil.rmtree will mess up symlinked directories.

BUG=
R=sigmund@google.com

Review URL: https://chromiumcodereview.appspot.com//680833003

+++ annotated_steps.py
@@ -160,7 +160,11 @@ def GetPackageCopy(bot_info):
   package_path = GetPackagePath(bot_info)
   copy_path = os.path.join(package_copy, bot_info.package_name)
   # Clean out old copy
-  shutil.rmtree(package_copy, ignore_errors=True)
+  if bot_info.system == 'windows':
+    args = ['cmd.exe', '/c', 'rmdir', '/q', '/s', package_copy]
+    RunProcess(args)
+  else:
+    shutil.rmtree(package_copy, ignore_errors=True)
   no_git = shutil.ignore_patterns('.git')
   shutil.copytree(package_path, copy_path, symlinks=False, ignore=no_git)
   return copy_path
